Add sin.erl with basic proof of concept for Sin.
I have a long ways to go before Sin is something that can be used
to do what it is I want it to do. The more I learn about Erlang,
the more I learn of what I should be doing. So this could take
a few iterations before I get it right.